+++ packages/x-charts-pro/src/ChartZoomSlider/internals/ChartAxisZoomSliderChartPreview.tsx
@@ -1,0 +1,11 @@
+import * as React from 'react';
+import { ScatterPreview } from './previews/ScatterPreview';
+
+export function ChartAxisZoomSliderChartPreview(props: {
+  x: number;
+  y: number;
+  height: number;
+  width: number;
+}) {
+  return <ScatterPreview {...props} />;
+}
